+++ changelog.md
@@ -17,6 +17,7 @@ __                                                                              
 __                                                                                                                 __
 ### Others
 * Camera Shake On Explosions
+* Tundra Bush Group
 __                                                                                                                 __
 ## Changes
 __                                                                                                                 __